fix(windows): compare GetExitCodeProcess result against STILL_ACTIVE as u32
PR #101 run 35379364786 (round 5): mux-core now fails on exactly one
line (two diagnostics, same line):

  win.rs:156  ok != 0 && exit_code == STILL_ACTIVE
  E0308 expected `u32`, found `i32` / E0277 can't compare u32 with i32

GetExitCodeProcess writes a DWORD (u32) but windows-sys 0.61 types
STILL_ACTIVE as NTSTATUS (i32; 259). Cast the constant, not the value,
and note why the cast is lossless (259 fits u32 trivially).

Signedness sweep of every other windows-sys constant comparison in the
workspace against the vendored windows-sys 0.61.2 definitions: WAIT_
OBJECT_0/WAIT_TIMEOUT are WAIT_EVENT (u32) and only ever compared with
WaitForSingleObject's u32 return; INVALID_HANDLE_VALUE is HANDLE and
compared with HANDLE; ReadFile/GetConsoleMode/GetCurrentConsoleFont/
SetConsoleCtrlHandler/CTRL_*_EVENT and the BOOL != 0 idioms all type-
check; mux-tui's win_console.rs, host_colors.rs windows probe,
graphics.rs cell-pixel probe and main.rs console handler were checked
item by item — no other mismatch found. This line is the round-5 delta.

Linux invariance: win.rs is #![cfg(windows)]; the file is invisible to
Linux/macOS compilation. Verified: cargo check -p mux-tui green.

+++ mux/crates/mux-core/src/win.rs
@@ -153,7 +153,10 @@ pub fn is_process_alive(pid: u32) -> bool {
         let mut exit_code: u32 = 0;
         let ok = GetExitCodeProcess(process, &mut exit_code);
         CloseHandle(process);
-        ok != 0 && exit_code == STILL_ACTIVE
+        // STILL_ACTIVE (259) is an NTSTATUS (i32) in windows-sys while
+        // GetExitCodeProcess writes u32; the value fits u32 trivially,
+        // so compare as u32.
+        ok != 0 && exit_code == STILL_ACTIVE as u32
     }
 }
